requirements.txt
discord.py
pre-commit.py